A lot of bug fixes and tweaks

+++ src/game.cpp
@@ -15,13 +15,13 @@ void draw_board(sf::RenderWindow &window) {
 
   sf::RectangleShape matrix_border(
       sf::Vector2f((CELL_SIZE + GAP) * COLUMNS + GAP * 3,
-                   (CELL_SIZE + GAP) * ROWS + GAP * 3));
+                   (CELL_SIZE + GAP) * (ROWS - 2) + GAP * 3));
   matrix_border.setPosition(LEFT_BORDER - GAP, TOP_BORDER - GAP);
   matrix_border.setFillColor(sf::Color(MIDNIGHT_BLUE));
   window.draw(matrix_border);
 
   sf::RectangleShape matrix_background(sf::Vector2f(
-      (CELL_SIZE + GAP) * COLUMNS - GAP, (CELL_SIZE + GAP) * ROWS - GAP));
+      (CELL_SIZE + GAP) * COLUMNS, (CELL_SIZE + GAP) * (ROWS - 2)));
   matrix_background.setPosition(LEFT_BORDER + GAP, TOP_BORDER + GAP);
   matrix_background.setFillColor(sf::Color(DARK_GRAY));
   window.draw(matrix_background);
@@ -30,26 +30,17 @@ void draw_board(sf::RenderWindow &window) {
 void draw_cells(sf::RenderWindow &window, matrixType &matrix,
                 TetrominoVariant &piece) {
 
-  sf::RectangleShape cell(sf::Vector2f(CELL_SIZE, CELL_SIZE));
-  cell.setFillColor(sf::Color(180, 50, 20));
+  sf::RectangleShape empty_cell(sf::Vector2f(CELL_SIZE, CELL_SIZE));
+  empty_cell.setFillColor(sf::Color(180, 50, 20));
 
   // Draw the piece with assigned colour
   sf::RectangleShape block(sf::Vector2f(CELL_SIZE, CELL_SIZE));
   auto piece_tag = std::visit([](auto &arg) { return arg.piece_tag; }, piece);
 
-  sf::RectangleShape background(sf::Vector2f(WINDOW_X, WINDOW_Y));
-  background.setFillColor(sf::Color(LIGHT_GRAY));
-  window.draw(background);
-
-  sf::RectangleShape matrix_background(sf::Vector2f(
-      (CELL_SIZE + GAP) * COLUMNS + GAP, (CELL_SIZE + GAP) * ROWS + GAP));
-  matrix_background.setPosition(LEFT_BORDER, TOP_BORDER);
-  matrix_background.setFillColor(sf::Color(DARK_GRAY));
-  window.draw(matrix_background);
-
   float x = GAP + LEFT_BORDER, y = GAP + TOP_BORDER;
 
-  for (int i = 0; i < ROWS; i++) {
+  // Pieces slowly reveal themselves
+  for (int i = 2; i < ROWS; i++) {
     for (int j = 0; j < COLUMNS; j++) {
 
       auto cell = matrix[i][j];
@@ -67,6 +58,7 @@ void draw_cells(sf::RenderWindow &window, matrixType &matrix,
         auto [r, g, b] = PIECE_COLOURS_MAP.at(piece_tag);
         block.setFillColor(sf::Color(r, g, b));
         block.setPosition(x, y);
+
         window.draw(block);
       }
       x += CELL_SIZE + GAP;
@@ -78,7 +70,8 @@ void draw_cells(sf::RenderWindow &window, matrixType &matrix,
 
 void handle_key_presses(sf::Event &ev, TetrominoVariant &piece,
                         pieceCoords &start_piece, coords &offset,
-                        matrixType &matrix, Score &score) {
+                        matrixType &matrix, Score &score, sf::Time &gameTick,
+                        sf::Clock &clock) {
   // Declaration here to prevent "jump bypasses variable initialisation" error
   bool valid_rotation = true;
   coords prev_offset;
@@ -90,8 +83,9 @@ void handle_key_presses(sf::Event &ev, TetrominoVariant &piece,
     break;
   case sf::Keyboard::S:
   case sf::Keyboard::Down:
-    movePiece(matrix, start_piece, 'd', offset);
-    score.tick();
+    if (movePiece(matrix, start_piece, 'd', offset)) {
+      score.tick();
+    }
     break;
   case sf::Keyboard::A:
   case sf::Keyboard::Left:
@@ -103,8 +97,11 @@ void handle_key_presses(sf::Event &ev, TetrominoVariant &piece,
       prev_offset = offset;
       movePiece(matrix, start_piece, 'd', offset);
       curr_offset = offset;
+      if (prev_offset != curr_offset) {
+        score.drop();
+      }
     } while (prev_offset != curr_offset);
-    score.drop();
+    seal_piece(piece, start_piece, offset, matrix);
     break;
     offset = curr_offset;
   case (sf::Keyboard::Q):
@@ -138,13 +135,7 @@ void handle_game_tick(matrixType &matrix, TetrominoVariant &piece,
 
   if (clock.getElapsedTime() > gameTick) { // game tick
     if (shouldSeal(matrix, start_piece, offset)) {
-      auto piece_type =
-          std::visit([](auto &arg) { return arg.piece_tag; }, piece);
-      set_piece_non_sealed(start_piece, offset, matrix, piece_type);
-      piece = std::move(choose_random(tetromino_piece_types));
-      start_piece = std::visit(
-          [](auto &arg) -> pieceCoords { return arg.getBlockCoords(); }, piece);
-      offset = std::make_tuple(COLUMNS / 2 - 2, 0);
+      seal_piece(piece, start_piece, offset, matrix);
     }
     clock.restart(); // Reset the clock
     movePiece(matrix, start_piece, 'd', offset);
@@ -152,6 +143,15 @@ void handle_game_tick(matrixType &matrix, TetrominoVariant &piece,
   }
 }
 
+void seal_piece(TetrominoVariant &piece, pieceCoords &start_piece,
+                coords &offset, matrixType &matrix) {
+  auto piece_type = std::visit([](auto &arg) { return arg.piece_tag; }, piece);
+  set_piece_non_sealed(start_piece, offset, matrix, piece_type);
+  piece = std::move(choose_random(tetromino_piece_types));
+  start_piece = std::visit(
+      [](auto &arg) -> pieceCoords { return arg.getBlockCoords(); }, piece);
+  offset = std::make_tuple(COLUMNS / 2 - 2, 0);
+}
 void set_piece_non_sealed(pieceCoords &start_piece, coords &offset,
                           matrixType &matrix, cell_type type) {
   for (auto [c_x, c_y] : start_piece) {
